feat: bottom-sheet menu preview button

+++ pages/timeline/[slug]/files/[id]/index.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import ListText from '~/components/ListText.vue'
-import { IFile } from '~/types/portfolio'
+import {IFile} from '~/types/portfolio'
 
 definePageMeta({
   title: 'File - Portfolio',
@@ -11,19 +11,19 @@ const route = useRoute()
 
 const router = useRouter()
 
-const { data, error } = await useAsyncData(
-  'timeline-data-file',
-  async () =>
-    await $fetch(
-      `/api/timeline/${route.params.slug}/files/${route.params.id}`,
-      {
-        method: 'GET',
-        headers: useRequestHeaders(['cookie'])
-      }
-    ),
-  {
-    transform: (res) => res.data.at(0) as IFile
-  }
+const {data, error} = await useAsyncData(
+    'timeline-data-file',
+    async () =>
+        await $fetch(
+            `/api/timeline/${route.params.slug}/files/${route.params.id}`,
+            {
+              method: 'GET',
+              headers: useRequestHeaders(['cookie'])
+            }
+        ),
+    {
+      transform: (res) => res.data.at(0) as IFile
+    }
 )
 
 if (!data.value) {
@@ -34,45 +34,49 @@ if (!data.value) {
   })
 }
 
+const isBusyDelete = ref<boolean>(false)
+
 const onDelete = async () => {
-  const { data, error } = await $fetch(
-    `/api/timeline/${route.params.slug}/files/${route.params.id}`,
-    {
-      method: 'DELETE',
-      headers: useRequestHeaders(['cookie'])
-    }
+  if (isBusyDelete.value) return
+  isBusyDelete.value = true
+  const {error} = await $fetch(
+      `/api/timeline/${route.params.slug}/files/${route.params.id}`,
+      {
+        method: 'DELETE',
+        headers: useRequestHeaders(['cookie'])
+      }
   )
+  isBusyDelete.value = false
 
   if (error) {
     console.error(error)
+    return
   }
 
-  if (data) {
-    await router.push(`/timeline/${route.params.slug}`)
-  }
+  await router.push(`/timeline/${route.params.slug}`)
 }
 </script>
 <template>
   <PageContainer>
     <div
-      v-if="data"
-      class="relative flex h-full w-full flex-col gap-4 overflow-hidden p-4"
+        v-if="data"
+        class="relative flex h-full w-full flex-col gap-4 overflow-hidden p-4"
     >
       <div class="flex">
         <div class="md:max-h-1/2">
-          <NuxtImg :src="data.filepath"  class="rounded-md"/>
+          <NuxtImg :src="data.filepath" class="rounded-md"/>
         </div>
       </div>
       <!-- If is admin -->
       <div class="relative flex flex-col">
         <div class="flex gap-4">
           <BaseButton
-            :disabled="!data"
-            :loading="!data"
-            class=""
-            @click="onDelete"
+              :disabled="!data || isBusyDelete"
+              :loading="!data || isBusyDelete"
+              class=""
+              @click="onDelete"
           >
-            Delete
+            Delete file
           </BaseButton>
         </div>
         <ListText>
